refactor(styles): Rename Abstractions to Abstracts
