+++ styles/abstracts/funcs.ts
